backend/requirements.txt
fastapi
uvicorn[standard]
insightface
opencv-python-headless
numpy
python-multipart
google-api-python-client
google-auth
python-dotenv
tqdm
Pillow
onnxruntime
